cmdata and make_mat updated (#543)
* updated cmdata to generate h5 histograms by default

* multiple updated for HDF5 format

* updated requirements

* flake8

* updated env

* github action

* gha

* flake

* gth

* again

* updated conda

* conda

* conda

* conda

* conda

* conda

* conda macos

* clean stuff

* miniforge

* again

* again

* escape s+

* removed escaped characters

* codecheck updated

* cmdata regtest

+++ tools/cmdata/src/cmdata/indexing.hpp
@@ -6,7 +6,7 @@
 namespace cmdata::indexing
 {
 
-int n_bins(const double cut, const double factor = 4.0)
+int n_bins(const float cut, const float factor = 4.0)
 {
   return cut / (0.01 / factor);
 }
